Ticket Classifier/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
httpx==0.24.0
PyYAML==5.4.1
pydantic==1.10.11